Upgrade to Celery 4, start using stunnel
This still needs some work, notably SSL verification and to figure out
if there's a supported way to use SSL for the backend (I'm almost
certain there is not).

I'm committing this now because it's not the *worst* thing in the world,
and I need to get create working again before tomorrow.